promoted new implementation of theta sketch
