+++ theta/include/theta_comparators.hpp
