+++ theta/include/theta_helpers.hpp
