+++ theta/include/theta_intersection_base.hpp
