+++ theta/include/theta_intersection_base_impl.hpp
