+++ theta/include/theta_set_difference_base.hpp
