+++ theta/include/theta_set_difference_base_impl.hpp
