+++ theta/test/CMakeLists.txt
@@ -42,4 +42,5 @@ target_sources(theta_test
     theta_union_test.cpp
     theta_intersection_test.cpp
     theta_a_not_b_test.cpp
+    theta_jaccard_similarity_test.cpp
 )
